requirements_production.txt
# Production requirements for DDoS AI Detection System
# Core Django Framework
Django==5.1.4
django-simple-captcha==0.6.0

# Production WSGI server
gunicorn>=21.0.0

# Database drivers
pymysql>=1.0.0

# Deep Learning and Machine Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Data Visualization and Analysis
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
pyecharts>=2.0.0

# Network and Security (limited for cloud deployment)
requests>=2.25.0
urllib3>=1.26.0

# Database and ORM
sqlparse>=0.4.0

# Web Development
Pillow>=8.0.0
asgiref>=3.4.0

# System and Utilities
psutil>=5.8.0
python-dateutil>=2.8.0
pytz>=2021.1

# Additional dependencies for the application
six>=1.16.0
setuptools>=65.0.0
wheel>=0.37.0

# For better compatibility
packaging>=21.0
importlib-metadata>=4.0.0

# Note: Scapy is removed for cloud deployment as it requires root privileges
# and can cause issues on cloud platforms
